flask/requirements.txt
absl-py==0.8.0
astor==0.8.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.16.4
opencv-python
pandas==0.25.1
Pillow==6.1.0
protobuf==3.9.1
pyparsing==2.4.2
pytesseract==0.3.0
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
